feat: transaction

+++ server/src/controllers/foodsContoller.js
@@ -4,38 +4,49 @@ import Food from '../models/Food.js';
 import fs from 'fs';
 import path from 'path';
 
-export const getAllFoodController = async (req, res, next) => {
-  try {
-    const foods = await Food.find();
+export const getAllFoodController = asyncHandler(async (req, res) => {
+  const { search, category } = req.query;
 
-    if (foods.length === 0) {
-      return res.status(200).json({
-        success: false,
-        message: 'there are no foods yet',
-        result: [],
-      });
-    }
+  const filter = {};
 
-    res.status(200).json({
-      success: true,
-      message: 'Enjoy your foods',
-      total: foods.length,
-      result: foods,
-    });
-  } catch (err) {
-    next(err);
+  if (search) {
+    filter.name = { $regex: search, $options: 'i' };
   }
-};
+
+  if (category) {
+    filter.category = category;
+  }
+
+  const foods = await Food.find(filter);
+
+  if (foods.length === 0) {
+    return res.status(200).json({
+      success: false,
+      message: 'there are no foods yet',
+      result: [],
+    });
+  }
+
+  res.status(200).json({
+    success: true,
+    message: 'Enjoy your foods',
+    total: foods.length,
+    result: foods,
+  });
+});
 
 export const createFoodsController = asyncHandler(async (req, res, next) => {
-  const { name, description, price, category } = req.body;
+  const { name, description, price, category, stock } = req.body;
 
-  if (!name || !description || !price || !category) {
+  const deleteFile = () => {
     if (req.file) {
       const pathToDelete = path.join('uploads', req.file.filename);
-
       if (fs.existsSync(pathToDelete)) fs.unlinkSync(pathToDelete);
     }
+  };
+
+  if (!name || !description || !price || !category || stock === undefined) {
+    deleteFile();
 
     return res.status(400).json({
       success: false,
@@ -43,18 +54,23 @@ export const createFoodsController = asyncHandler(async (req, res, next) => {
     });
   }
 
+  const existingName = await Food.findOne({ name: name });
+
+  if (existingName) {
+    deleteFile();
+    return res
+      .status(409)
+      .json({ success: false, message: 'Food name already exist' });
+  }
+
   const categoryId = await Category.findById(category);
 
   if (!categoryId) {
-    if (req.file) {
-      const pathToDelete = path.join('uploads', req.file.filename);
-
-      if (fs.existsSync(pathToDelete)) fs.unlinkSync(pathToDelete);
-    }
+    deleteFile();
 
     return res.status(404).json({
       success: false,
-      messaage: 'categoryId not found, please provide a valid category ID',
+      message: 'categoryId not found, please provide a valid category ID',
     });
   }
 
@@ -64,6 +80,7 @@ export const createFoodsController = asyncHandler(async (req, res, next) => {
     image: req.file ? req.file.filename : undefined,
     price,
     category,
+    stock,
   });
 
   if (!result) {
@@ -77,37 +94,54 @@ export const createFoodsController = asyncHandler(async (req, res, next) => {
   });
 });
 
-export const updateFoodController = asyncHandler(async (req, res, next) => {
+export const updateFoodController = asyncHandler(async (req, res) => {
   const { id } = req.params;
-  const { name, description, price, category } = req.body;
+  const { name, description, price, category, stock } = req.body;
+
   const newImage = req.file ? req.file.filename : undefined;
+
+  const deleteFile = () => {
+    if (req.file) {
+      const pathToDelete = path.join('uploads', req.file.filename);
+      if (fs.existsSync(pathToDelete)) fs.unlinkSync(pathToDelete);
+    }
+  };
+
+  if (category) {
+    const categoryExist = await Category.findById(category);
+
+    if (!categoryExist) {
+      deleteFile();
+
+      return res.status(404).json({
+        success: false,
+        message: 'categoryId not found, please provide a valid category ID',
+      });
+    }
+  }
 
   const existingFood = await Food.findById(id);
 
   if (!existingFood) {
-    if (newImage) {
-      const pathToDelete = path.join('uploads', newImage);
-      if (fs.existsSync(pathToDelete)) fs.unlinkSync(pathToDelete);
-    }
+    deleteFile();
 
     return res.status(404).json({ success: false, message: 'food not found' });
   }
 
-  const updateData = { name, description, price, category };
+  const updateData = {
+    name,
+    description,
+    price,
+    category,
+    stock,
+    ...(newImage && { image: newImage }),
+  };
 
-  if (newImage) {
-    updateData.image = newImage;
+  if (newImage && existingFood.image) {
+    const oldImagePath = path.join('uploads', existingFood.image);
 
-    if (existingFood.image) {
-      const oldImagePath = path.join('uploads', existingFood.image);
-
-      if (fs.existsSync(oldImagePath)) {
-        try {
-          fs.unlinkSync(oldImagePath);
-        } catch (err) {
-          next(err);
-        }
-      }
+    if (fs.existsSync(oldImagePath)) {
+      fs.unlinkSync(oldImagePath);
     }
   }
 
@@ -140,18 +174,14 @@ export const deleteFoodController = asyncHandler(async (req, res, next) => {
     const pathToDelete = path.join('uploads', food.image);
 
     if (fs.existsSync(pathToDelete)) {
-      try {
-        fs.unlinkSync(pathToDelete);
-      } catch (err) {
-        next(err);
-      }
+      fs.unlinkSync(pathToDelete);
     }
   }
 
-  await Food.findByIdAndDelete(id);
+  await food.deleteOne();
 
   res.status(200).json({
     success: true,
-    message: 'Food deleted successfully',
+    message: 'Food and associated image deleted successfully',
   });
 });
